Update Dockerfile
Use case:
- Running esphome inside docker behind a reverse proxy without exposing 6052 to the host interface

Why:
When you don't expose 6052 in the dockerfile, you cannot use a reverse proxy inside docker-compose. It only works when you explicitly bind 6052 to a port of the host interface. (with docker run --expose for example)  The whole reason for using a reverse proxy is to not have a ton of ports exposed. 

Note:
"EXPOSE" in a dockerfile is not the same as --expose. When EXPOSE is used, but the image isn't started with --expose, the port is only exposed to the internal docker network.